count_service/server/requirements.txt
grpcio
grpcio-tools
redis
grpcio-health-checking